+++ src/wheel_inspect/__init__.py
@@ -14,8 +14,9 @@ or from the command line with the ``wheel2json`` command.
 Visit <https://github.com/jwodder/wheel-inspect> for more information.
 """
 
-from .schema   import SCHEMA
-from .wheelcls import Wheel, inspect_wheel
+from wheel_filename import ParsedWheelFilename, parse_wheel_filename
+from .schema        import SCHEMA
+from .wheelcls      import Wheel, inspect_wheel
 
 __version__      = '1.5.0.dev1'
 __author__       = 'John Thorvald Wodder II'
@@ -24,7 +25,9 @@ __license__      = 'MIT'
 __url__          = 'https://github.com/jwodder/wheel-inspect'
 
 __all__ = [
+    'ParsedWheelFilename',
     'SCHEMA',
     'Wheel',
     'inspect_wheel',
+    'parse_wheel_filename',
 ]