+++ src/wheel_inspect/__init__.py
@@ -18,7 +18,7 @@ from wheel_filename import ParsedWheelFilename, parse_wheel_filename
 from .schema        import SCHEMA
 from .wheelcls      import Wheel, inspect_wheel
 
-__version__      = '1.5.0.dev1'
+__version__      = '1.5.0'
 __author__       = 'John Thorvald Wodder II'
 __author_email__ = 'wheel-inspect@varonathe.org'
 __license__      = 'MIT'